% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{.calculate_needed_pops}
\alias{.calculate_needed_pops}
\title{Calculate needed populations}
\usage{
.calculate_needed_pops(Fst, nmax = 20)
}
\arguments{
\item{Fst}{FST value.}

\item{nmax}{Maximum number of populations to evaluate.}
}
\value{
A data frame with population number and retention probability.
}
\description{
Internal helper that calculates the cumulative probability of retaining rare
alleles as the number of populations increases.
}
\keyword{internal}
