% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{make_data_preview}
\alias{make_data_preview}
\title{Create a limited data preview}
\usage{
make_data_preview(dta, max_rows = 50, max_cols = 100)
}
\arguments{
\item{dta}{A data frame.}

\item{max_rows}{Maximum number of rows to include.}

\item{max_cols}{Maximum number of columns to include.}
}
\value{
A data frame subset.
}
\description{
Returns the first rows and columns of a data frame for display in the Shiny
application.
}
