% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrguc-package.R
\docType{package}
\name{rrguc-package}
\alias{rrguc}
\alias{rrguc-package}
\title{rrguc: Genetic Conservation Shiny Application}
\description{
Tools and a Shiny application for genetic conservation analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dinilu/rrguc}
  \item Report bugs at \url{https://github.com/dinilu/rrguc/issues}
}

}
\author{
\strong{Maintainer}: Diego Nieto Lugilde \email{bv2nilud@uco}

Authors:
\itemize{
  \item Paula Morales Sandoica
  \item Daniel Romera Romera
}

}
\keyword{internal}
