% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rrguc_server}
\alias{rrguc_server}
\title{Shiny application server}
\usage{
rrguc_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
No return value. Called for side effects within Shiny.
}
\description{
Defines the server logic of the rrguc Shiny application.
}
