% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the rrguc Shiny application}
\usage{
run_app()
}
\value{
No return value. This function launches a Shiny app.
}
\description{
Launches the Shiny application included in the package.
}
