% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{select_alleles}
\alias{select_alleles}
\title{Select low-frequency alleles}
\usage{
select_alleles(mat, pops, allele_perc = 0.1, pop_perc = 0.1)
}
\arguments{
\item{mat}{Numeric matrix or data frame of genetic data.}

\item{pops}{Population identifier for each row of \code{mat}.}

\item{allele_perc}{Maximum overall allelic frequency.}

\item{pop_perc}{Maximum frequency across populations.}
}
\value{
A character vector with selected allele names.
}
\description{
Selects alleles according to maximum individual allele frequency and maximum
population frequency thresholds.
}
